Fixed a bug in zernike_compute_EyEx
Thanks to Bing Dong (icybird@qq.com) for letting me know about this.

There were some errors in the computation of the mean gradients of the
Zernike modes. When the centre of the pupil lies within a subaperture
the limits of the integral should be treated differently. Also the
formula for the limits of theta should be with acos instead of atan.
After this commit the mean gradients of tip and tilt are constant as
should be expected.